backend/requirements.txt
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1
mysql-connector-python==9.0.0
mysqlclient==2.2.4
networkx==3.3
numpy==2.0.0
packaging==24.1
pillow==10.4.0
pycocotools==2.0.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
six==1.16.0
SQLAlchemy==2.0.31
tifffile==2024.7.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
